requirements.txt
flask
gunicorn
requests
python-dotenv
scikit-learn
numpy
pandas
psycopg2-binary
